Removed file object convertors and moved all those mappers to related models
