+++ app/src/main/java/com/developersbreach/composeactors/data/model/Actor.kt
@@ -2,6 +2,7 @@ package com.developersbreach.composeactors.data.model
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
+import com.developersbreach.composeactors.data.database.entity.FavoriteActorsEntity
 
 @Immutable
 data class Actor(
@@ -9,3 +10,21 @@ data class Actor(
     val actorName: String,
     val profileUrl: String
 )
+
+fun Actor.actorAsDatabaseModel(): FavoriteActorsEntity {
+    return FavoriteActorsEntity(
+        actorId = this.actorId,
+        actorName = this.actorName,
+        actorProfileUrl = this.profileUrl
+    )
+}
+
+fun List<FavoriteActorsEntity>.actorAsDomainModel(): List<Actor> {
+    return map {
+        Actor(
+            actorId = it.actorId,
+            actorName = it.actorName,
+            profileUrl = it.actorProfileUrl
+        )
+    }
+}
